requirements.txt
pygame>=2.5.2
paho-mqtt>=2.0.0
